+++ src/user-cpp/share/module_callsvc.cpp
@@ -1,4 +1,5 @@
 #include <stdlib.h>
+#include <unistd.h>
 #include <string.h>
 
 #include <string>
@@ -39,7 +40,7 @@ void module_release()
 }
 
 static
-size_t module_unpack(skullcpp::Txn& txn, const void* data, size_t data_sz)
+ssize_t module_unpack(skullcpp::Txn& txn, const void* data, size_t data_sz)
 {
     moduleMetrics.request.inc(1);
 
@@ -49,7 +50,7 @@ size_t module_unpack(skullcpp::Txn& txn, const void* data, size_t data_sz)
     // deserialize data to transcation data
     auto& example = (skull::workflow::example&)txn.data();
     example.set_data(data, data_sz);
-    return data_sz;
+    return (ssize_t)data_sz;
 }
 
 static
